Work in progress -- Error including cstdint

+++ app/src/main/jni/tutorial2_jnipart.cpp
@@ -2,7 +2,6 @@
 #include <opencv2/core/core.hpp>
 #include <opencv2/imgproc/imgproc.hpp>
 #include <opencv2/features2d/features2d.hpp>
-#include <vector>
 
 using namespace std;
 using namespace cv;
@@ -21,7 +20,7 @@ JNIEXPORT void JNICALL Java_com_jnardari_opencv_1androidsamples_activities_Tutor
     for( unsigned int i = 0; i < v.size(); i++ )
     {
         const KeyPoint& kp = v[i];
-        circle(mRgb, Point(kp.pt.x, kp.pt.y), 10, Scalar(255,0,0,255));
+        circle(mRgb, Point((int)kp.pt.x, (int)kp.pt.y), 10, Scalar(255,0,0,255));
     }
 }
 }
